lib/mito/geometry: split coordinate-type specific implementations to subdirectories
