+++ lib/mito/geometry/cartesian.h
@@ -1,0 +1,17 @@
+// -*- c++ -*-
+//
+// Copyright (c) 2020-2024, the MiTo Authors, all rights reserved
+//
+
+// code guard
+#if !defined(mito_geometry_cartesian_h)
+#define mito_geometry_cartesian_h
+
+
+// publish the interface
+#include "cartesian/public.h"
+
+
+#endif
+
+// end of file
