+++ lib/mito/geometry/cartesian/algebra.h
