+++ lib/mito/geometry/polar/public.h
@@ -1,0 +1,21 @@
+// -*- c++ -*-
+//
+// Copyright (c) 2020-2024, the MiTo Authors, all rights reserved
+//
+
+// code guard
+#if !defined(mito_geometry_polar_public_h)
+#define mito_geometry_polar_public_h
+
+
+// get the coordinate basis
+#include "CoordinateBasis.h"
+// get the metric
+#include "metric.h"
+// get the algebra
+#include "algebra.h"
+
+
+#endif
+
+// end of file
